requirements.txt
apache-airflow[postgres]
dvc[gs]
flask
scikit-learn
scipy
pandas
numpy
apache-airflow-providers-fab
python-dotenv